Started building an application for this project.
Backend and API are essentially done, I used Django and DRF to build a endpoint which displays stock returns based on the ticker the user inputs from the frontend.

This is NOT production code and has some major security vulnerabilities. However this is just to show the basic methodology of building a full stack application.

The next step is to build the React + Tailwind frontend, to display a plotly.js histogram based on what the user inputs.